Initial OSA zuul v3 role jobs
This patch implements an initial set of jobs intended to match
the current job execution method. It does not intend to improve
how the jobs are executed - only to replicate what is currently
in openstack-infra/openstack-zuul-jobs and provide the platform
to iterate on.

Change-Id: I653f7b73ded5a2648c2d519136e261d0ea81c1c1
(cherry picked from commit b057efe2c55633e487db16802ffedf66d5d351f9)