Clean up role tests
Now that we're using the general templates, we can slim down
the role test definitions. We can also remove tests which are
not being watched, or which are fundamentally broken. With
this we can also remove unused scripts/plays.

We do the following:

1. We revert the use of project templates to using job config
   directly, given that nothing other than Ubuntu Xenial/Bionic
   is currently working.

Change-Id: I37a9f2af33d47a8edd5788d1504bb2c956356235
(cherry picked from commit bf83323e97483f42315664c566d8806678eb59e7)